tests for canonicalize
