+++ unittests/operators/manipulation.cpp
