+++ unittests/operators/product_op.cpp
@@ -14,6 +14,32 @@
 
 TEST(OperatorExpressions, checkProductOperatorBasics) {
 
+  // checking some utility functions
+  {
+    for (auto id_target = 0; id_target < 4; ++id_target) {
+      cudaq::spin_op_term op;
+      cudaq::spin_op_term expected;
+      for (std::size_t target = 0; target < 4; ++target) {
+        if (target == id_target) op *= cudaq::spin_op::i(target);
+        else if (target & 2) {
+          op *= cudaq::spin_op::z(target);
+          expected *= cudaq::spin_op::z(target);
+        } else if (target & 1) {
+          op *= cudaq::spin_op::x(target);
+          expected *= cudaq::spin_op::x(target);
+        } else {
+          op *= cudaq::spin_op::y(target);
+          expected *= cudaq::spin_op::y(target);
+        }
+      }
+      ASSERT_NE(op, expected);
+      op.canonicalize();
+      ASSERT_EQ(op, expected);
+      ASSERT_EQ(op.degrees(), expected.degrees());
+      ASSERT_EQ(op.to_matrix(), expected.to_matrix());
+    }
+  }
+
   // checking some constructors
   {
     cudaq::product_op<cudaq::matrix_handler> ids(2, 5);
